requirements.txt
chromadb
google-genai
python-dotenv
fastapi
uvicorn
